backend/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# feature_extraction/generate_emotion_feature.py: 10
Pillow == 5.1.0

# experiment/test.py: 4
# experiment/train.py: 9
asposebarcode == 1.0.0

# experiment/test.py: 4
# experiment/train.py: 9
asposestorage == 1.0.2

# experiment/test.py: 4
# experiment/train.py: 9
automium-web == 0.1.1

# experiment/test.py: 5
# experiment/train.py: 10
data == 0.4

# experiment/models.py: 17
# experiment/test.py: 10
# experiment/train.py: 15
functions == 0.7.0

# experiment/test.py: 4
# experiment/train.py: 9
hematopy == 0.0.1.dev6

# experiment/test.py: 4
# experiment/train.py: 9
indras-net == 2.0.6

# experiment/batch_renorm.py: 3,4
# experiment/data.py: 18
# experiment/functions.py: 2
# experiment/models.py: 12,13
# experiment/test.py: 1,2,3
# experiment/train.py: 6,7,8
# feature_extraction/features_from_file.py: 20,21,22
# feature_extraction/generate_body_visual.py: 4,5,16
# feature_extraction/generate_emotion_feature.py: 12,13
# feature_extraction/generate_face_visual.py: 4,5,15
keras == 2.2.4

# feature_extraction/features_from_file.py: 18
# feature_extraction/generate_body_visual.py: 6
# feature_extraction/generate_face_visual.py: 6
keras-vggface == 0.5

# experiment/utils.py: 12
matplotlib

# feature_extraction/generate_word_feature.py: 9,10,12
nltk == 3.4.1

# calculateEvaluationCCC.py: 2,10
# data_preparation/create_batches.py: 4
# experiment/data.py: 13
# experiment/fix_pkl.py: 4
# experiment/functions.py: 1
# experiment/test.py: 11
# experiment/train.py: 16
# experiment/utils.py: 11
# feature_extraction/extract_audio_files.py: 11
# feature_extraction/features_from_file.py: 11
# feature_extraction/generate_audio_feature.py: 26
# feature_extraction/generate_body_all.py: 4
# feature_extraction/generate_body_feature.py: 4
# feature_extraction/generate_body_visual.py: 11
# feature_extraction/generate_emotion_feature.py: 4
# feature_extraction/generate_face_feature.py: 9
# feature_extraction/generate_face_visual.py: 10
numpy == 1.16.2

# feature_extraction/features_from_file.py: 3
# feature_extraction/generate_body_all.py: 3
# feature_extraction/generate_body_feature.py: 3
# feature_extraction/generate_emotion_feature.py: 3
opencv-contrib-python == 4.1.0.25

# feature_extraction/features_from_file.py: 3
# feature_extraction/generate_body_all.py: 3
# feature_extraction/generate_body_feature.py: 3
# feature_extraction/generate_emotion_feature.py: 3
opencv-contrib-python-headless == 4.1.0.25

# feature_extraction/features_from_file.py: 3
# feature_extraction/generate_body_all.py: 3
# feature_extraction/generate_body_feature.py: 3
# feature_extraction/generate_emotion_feature.py: 3
opencv-python == 4.1.0.25

# feature_extraction/features_from_file.py: 3
# feature_extraction/generate_body_all.py: 3
# feature_extraction/generate_body_feature.py: 3
# feature_extraction/generate_emotion_feature.py: 3
# opencv-python-aarch64 == 3.3.0.1

# feature_extraction/features_from_file.py: 3
# feature_extraction/generate_body_all.py: 3
# feature_extraction/generate_body_feature.py: 3
# feature_extraction/generate_emotion_feature.py: 3
opencv-python-headless == 4.1.0.25

# calculateEvaluationCCC.py: 11
# data_preparation/create_batches.py: 6
# experiment/export_evaluations.py: 5
# experiment/export_results.py: 5
# experiment/utils.py: 13
# feature_extraction/features_from_file.py: 12
# feature_extraction/generate_body_visual.py: 14
# feature_extraction/generate_face_feature.py: 17
# feature_extraction/generate_face_visual.py: 13
pandas == 0.24.2

# experiment/test.py: 4
# experiment/train.py: 9
#pynotion

# experiment/test.py: 4
# experiment/train.py: 9
pyskyscanner == 1.0.1

# experiment/test.py: 4
# experiment/train.py: 9
roomai == 0.1.16

# calculateEvaluationCCC.py: 15,19
# experiment/data.py: 20
# experiment/utils.py: 17
# feature_extraction/features_from_file.py: 24
scikit-learn == 0.20.3

# calculateEvaluationCCC.py: 15,19
# experiment/data.py: 20
# experiment/utils.py: 17
# feature_extraction/features_from_file.py: 24
scikit-learn-runnr == 0.18.dev1

# calculateEvaluationCCC.py: 9
scipy == 1.2.1

# experiment/test.py: 4
# experiment/train.py: 9
#solar-radiation-model

# experiment/test.py: 4
# experiment/train.py: 9
#swiftcodegen

# experiment/batch_renorm.py: 12
# experiment/test.py: 6
# experiment/train.py: 11
# feature_extraction/generate_emotion_feature.py: 14
tensorflow == 1.13.1

# experiment/batch_renorm.py: 15
theano == 1.0.4

# feature_extraction/extract_audio_files.py: 12
# feature_extraction/features_from_file.py: 13
# feature_extraction/generate_audio_feature.py: 25,27
# feature_extraction/generate_body_all.py: 10
# feature_extraction/generate_body_feature.py: 10
# feature_extraction/generate_body_visual.py: 7
# feature_extraction/generate_emotion_feature.py: 11
# feature_extraction/generate_face_feature.py: 14
# feature_extraction/generate_face_visual.py: 7
# feature_extraction/generate_word_feature.py: 16
tqdm == 4.31.1

# experiment/test.py: 4
# experiment/train.py: 9
transferflow == 0.1.8

# experiment/data.py: 19
# experiment/test.py: 17
# experiment/train.py: 20
utils == 0.9.0

# experiment/test.py: 4
# experiment/train.py: 9
visvmtagger
